+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Press ⏎ on a displayed answer to copy it to the clipboard
 - Improved icon
 - Changed keyword to "cla"
 - Improved styling of output